services/scraper-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
playwright
asyncio
boto3
httpx
